requirements.txt
streamlit
pandas
openpyxl
python-dotenv
folium
streamlit-folium
googlemaps
google-cloud-aiplatform